pull crypto wallet row into its own component file to clean code and fix some react nesting issues
