+++ client/src/components/Home/ProfileDrawer/CryptoWallet.tsx
@@ -1,57 +1,12 @@
-import React from "react";
-import Collapse from "@mui/material/Collapse";
-import IconButton from "@mui/material/IconButton";
 import Table from "@mui/material/Table";
 import TableBody from "@mui/material/TableBody";
 import TableCell from "@mui/material/TableCell";
 import TableContainer from "@mui/material/TableContainer";
 import TableHead from "@mui/material/TableHead";
 import TableRow from "@mui/material/TableRow";
-import Typography from "@mui/material/Typography";
 import Paper from "@mui/material/Paper";
-import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
-import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
-import { Box } from "@mui/material";
 import { useUserContext } from "../../../hooks/useUserContext";
-
-interface RowType {
-  row: { crypto: string; quantity: number; principle: number };
-}
-
-function Row({ row }: RowType) {
-  const [open, setOpen] = React.useState(false);
-
-  return (
-    <React.Fragment>
-      <TableRow sx={{ "& > *": { borderBottom: "unset" } }}>
-        <TableCell>
-          <IconButton
-            aria-label="expand row"
-            size="small"
-            onClick={() => setOpen(!open)}
-          >
-            {open ? <KeyboardArrowUpIcon /> : <KeyboardArrowDownIcon />}
-          </IconButton>
-        </TableCell>
-        <TableCell component="th" scope="row">
-          {row.crypto}
-        </TableCell>
-        <TableCell align="right">{row.quantity}</TableCell>
-      </TableRow>
-      <TableRow>
-        <TableCell style={{ paddingBottom: 0, paddingTop: 0 }} colSpan={6}>
-          <Collapse in={open} timeout="auto" unmountOnExit>
-            <Box sx={{ margin: 1 }}>
-              <Typography variant="h6" gutterBottom component="div">
-                Value
-              </Typography>
-            </Box>
-          </Collapse>
-        </TableCell>
-      </TableRow>
-    </React.Fragment>
-  );
-}
+import { WalletRow } from "./WalletRow";
 
 export const CryptoWallet = () => {
   const { user } = useUserContext();
@@ -67,7 +22,7 @@ export const CryptoWallet = () => {
         </TableHead>
         <TableBody>
           {user?.portfolio.map((row) => (
-            <Row key={row.crypto} row={row} />
+            <WalletRow key={row.crypto} row={row} />
           ))}
         </TableBody>
       </Table>
